+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 Changes to the HexSet distribution. The project follows
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## 0.48.2
 
 ### Fixed